+++ src/components/Form/Col/Col.tsx
@@ -20,7 +20,7 @@ import FormFieldRaw from "@root/components/Field/FormFieldRaw";
 import FormFieldTable from "@root/components/Field/FormFieldTable";
 import FormFieldText from "@root/components/Field/FormFieldText";
 import FormFieldTextEditor from "@root/components/Field/FormFieldTextEditor";
-import FormFieldToggleSwitch from "@root/forms/FormFieldToggleSwitch";
+import FormFieldToggleSwitch from "@root/components/Field/FormFieldToggleSwitch";
 import FormFieldUpload from "@root/forms/FormFieldUpload";
 
 import evaluateShow from "@root/utils/show/evaluateShow";
